requirements.txt
Django==4.2
djangorestframework
django-crispy-forms
django-ckeditor
drf-yasg==1.20.0
django-cors-headers
numpy
depthai
opencv-python

